+++ R/bss.R
@@ -5,6 +5,7 @@
 #' @param method see \code{\link{train}}
 #' @param metric see \code{\link{train}}
 #' @param maximize see \code{\link{train}}
+#' @param globalval Logical. Should models be evaluated based on 'global' performance? See \code{\link{global_validation}}
 #' @param trControl see \code{\link{train}}
 #' @param tuneLength see \code{\link{train}}
 #' @param tuneGrid see \code{\link{train}}
@@ -18,7 +19,7 @@
 #' performance. It also contains "perf_all" that gives the performance of all model runs.
 #' @details bss is an alternative to \code{\link{ffs}} and ideal if the training
 #' set is small. Models are iteratively fitted using all different combinations
-#' of predictor variables. Hence, 2^X models are calculated. Dont try running bss
+#' of predictor variables. Hence, 2^X models are calculated. Don't try running bss
 #' on very large datasets because the computation time is much higher compared to
 #' \code{\link{ffs}}.
 #'
@@ -26,9 +27,9 @@
 #' on parallel processing of carets train functions for details.
 #'
 #'
-#' @note This validation is particulary suitable for spatial
-#' leave-location-out cross validations where variable selection
-#' MUST be based on the performance of the model on the hold out station.
+#' @note This variable selection is particularly suitable for spatial
+#' cross validations where variable selection
+#' MUST be based on the performance of the model for predicting new spatial units.
 #' Note that bss is very slow since all combinations of variables are tested.
 #' A more time efficient alternative is the forward feature selection (\code{\link{ffs}})
 #'  (\code{\link{ffs}}).
@@ -49,6 +50,7 @@ bss <- function (predictors,
                  method = "rf",
                  metric = ifelse(is.factor(response), "Accuracy", "RMSE"),
                  maximize = ifelse(metric == "RMSE", FALSE, TRUE),
+                 globalval=FALSE,
                  trControl = caret::trainControl(),
                  tuneLength = 3,
                  tuneGrid = NULL,
@@ -56,7 +58,8 @@ bss <- function (predictors,
                  verbose=TRUE,
                  ...){
   trControl$returnResamp <- "final"
-  if(class(response)=="character"){
+  trControl$savePredictions <- "final"
+  if(inherits(response,"character")){
     response <- factor(response)
     if(metric=="RMSE"){
       metric <- "Accuracy"
@@ -83,14 +86,25 @@ bss <- function (predictors,
                    response,method=method,trControl=trControl,
                    tuneLength=tuneLength,
                    tuneGrid=tuneGrid,...)
-    actmodelperf <- evalfunc(model$results[,names(model$results)==metric])
+
+
+    if (globalval){
+      perf_stats <- global_validation(model)[names(global_validation(model))==metric]
+    }else{
+      perf_stats <- model$results[,names(model$results)==metric]
+    }
+    actmodelperf <- evalfunc(perf_stats)
+
     actmodelperfSE <- se(
       sapply(unique(model$resample$Resample),
              FUN=function(x){mean(model$resample[model$resample$Resample==x,
                                                  metric],na.rm=TRUE)}))
+    bestmodelperfSE <- actmodelperfSE
     if (i == 1){
       bestmodelperf <- actmodelperf
-      bestmodelperfSE <- actmodelperfSE
+      if(globalval){
+        bestmodelperfSE <- NA
+      }
       bestmodel <- model
     } else{
       if (isBetter(actmodelperf,bestmodelperf,maximization=maximize)){
@@ -107,11 +121,20 @@ bss <- function (predictors,
                  2^n-(n+1) - acc))
     }
   }
-  if (maximize){
-    selectedvars_perf <- max(bestmodel$results[,metric])
-  } else{
-    selectedvars_perf <- min(bestmodel$results[,metric])
+
+
+
+  if (globalval){
+    selectedvars_perf <- global_validation(bestmodel)[names(global_validation(bestmodel))==metric]
+  }else{
+    if (maximize){
+      selectedvars_perf <-max(bestmodel$results[,metric])
+    }else{
+      selectedvars_perf <- min(bestmodel$results[,metric])
+    }
   }
+
+
   bestmodel$selectedvars <- bestmodel$finalModel$xNames
   bestmodel$selectedvars_perf <- selectedvars_perf
   bestmodel$perf_all <- perf_all