+++ NEWS.md
@@ -1,3 +1,15 @@
+# `CAST` 0.6.0
+* new feature: 
+  * trainDI allows to calculate the DI of the training dataset separately from the aoa function
+  * plot and print functions for the AOA
+  * function to plot nearest neighbor distance distributions in geographic and feature space
+  * function global_validation added
+* modifications
+  * extensive restructuring of the AOA function
+  * ffs and bss can be used with global_validation
+* bugfix:
+  * error in manual assignment of weights fixed
+
 # `CAST` 0.5.1
 * resolved dependence on package "GSIF" which was removed from the CRAN repository